+++ app/src/main/java/ru/gaket/themoviedb/presentation/moviedetails/viewmodel/MovieDetailsState.kt
@@ -1,3 +1,3 @@
 package ru.gaket.themoviedb.presentation.moviedetails.viewmodel
 
-data class MovieDetailsState(val movieId: Int, val title: String)
+data class MovieDetailsState(val movieId: Long, val title: String)
